Moved Spectrometer and Constants to own file
